cls-lua: generalize Lua VM so we can use it in the MDS
