+++ src/cls/Makefile-server.am
@@ -82,8 +82,9 @@ libcls_journal_la_LDFLAGS = ${AM_LDFLAGS} -module -avoid-version -shared -export
 radoslib_LTLIBRARIES += libcls_journal.la
 
 libcls_lua_la_SOURCES = \
-	cls/lua/cls_lua.cc \
-	cls/lua/lua_bufferlist.cc
+	cls/lua/cls_lua_rados.cc \
+        lua_core/lua_core.cc \
+	lua_core/lua_bufferlist.cc
 libcls_lua_la_CPPFLAGS = $(LIBLUA_CPPFLAGS) $(AM_CPPFLAGS)
 libcls_lua_la_LIBADD = $(LIBLUA)
 libcls_lua_la_LDFLAGS = ${AM_LDFLAGS} -module -avoid-version -shared -export-symbols-regex '.*__cls_.*'
